Refactor server logic/data from host.
