+++ Server.Core/Models/DbEntityConfiguration.cs
@@ -1,13 +1,13 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Quibble.Shared.Entities;
+using Quibble.Common.Entities;
 
-namespace Quibble.Server.Data.Models;
+namespace Quibble.Server.Core.Models;
 
 public class DbEntityConfiguration<TEntity> : IEntityTypeConfiguration<TEntity> where TEntity : class, IEntity
 {
-    public virtual void Configure(EntityTypeBuilder<TEntity> builder)
-    {
-        builder.HasKey(entity => entity.Id);
-    }
+	public virtual void Configure(EntityTypeBuilder<TEntity> builder)
+	{
+		builder.HasKey(entity => entity.Id);
+	}
 }
