+++ Server.Core/Sync/IExecutionContext.cs
@@ -1,0 +1,9 @@
+﻿namespace Quibble.Server.Core.Sync;
+
+public interface IExecutionContext
+{
+	public Guid UserId { get; }
+	public Guid QuizId { get; }
+
+	public ISyncedUsers Users { get; }
+}
