--- Server/Data/AppRole.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Quibble.Shared.Entities;
-
-namespace Quibble.Server.Data;
-
-public class AppRole : IdentityRole<Guid>, IEntity
-{
-}
