+++ Server/Extensions/EnumerableExtensions.cs
@@ -2,12 +2,12 @@
 
 public static class EnumerableExtensions
 {
-    public static bool None<TSource>(this IEnumerable<TSource> source, Func<TSource, bool> predicate)
-    {
-        foreach (TSource element in source)
-            if (predicate(element))
-                return false;
+	public static bool None<TSource>(this IEnumerable<TSource> source, Func<TSource, bool> predicate)
+	{
+		foreach (TSource element in source)
+			if (predicate(element))
+				return false;
 
-        return true;
-    }
+		return true;
+	}
 }
