+++ Server/Program.cs
@@ -2,10 +2,10 @@
 
 public static class Program
 {
-    public static void Main(string[] args) =>
-        CreateHostBuilder(args).Build().Run();
+	public static void Main(string[] args) =>
+		CreateHostBuilder(args).Build().Run();
 
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-        Host.CreateDefaultBuilder(args)
-            .ConfigureWebHostDefaults(webBuilder => webBuilder.UseStartup<Startup>());
+	public static IHostBuilder CreateHostBuilder(string[] args) =>
+		Host.CreateDefaultBuilder(args)
+			.ConfigureWebHostDefaults(webBuilder => webBuilder.UseStartup<Startup>());
 }
