+++ Server/Services/EmailSender/DebugToLogEmailSender.cs
@@ -2,16 +2,16 @@
 
 public class DebugToLogEmailSender : IEmailSender
 {
-    private readonly ILogger<DebugToLogEmailSender> _logger;
+	private readonly ILogger<DebugToLogEmailSender> _logger;
 
-    public DebugToLogEmailSender(ILogger<DebugToLogEmailSender> logger)
-    {
-        _logger = logger;
-    }
+	public DebugToLogEmailSender(ILogger<DebugToLogEmailSender> logger)
+	{
+		_logger = logger;
+	}
 
-    public Task SendAsync(string address, string content)
-    {
-        _logger.LogInformation("{Address}: {Content}", address, content);
-        return Task.CompletedTask;
-    }
+	public Task SendAsync(string address, string content)
+	{
+		_logger.LogInformation("{Address}: {Content}", address, content);
+		return Task.CompletedTask;
+	}
 }
